Adjusted tick sync correction algorithm to not fire adjustment for lag spikes.

+++ Source/Client/Game/Private/RenderSystem.cpp
@@ -26,7 +26,7 @@ void RenderSystem::render(double deltaSeconds)
         renderer.Clear();
 
         // How far we are between game ticks in decimal percent.
-        float alpha = game.getAccumulatedTime() / GAME_TICK_INTERVAL_S;
+        double alpha = game.getAccumulatedTime() / GAME_TICK_INTERVAL_S;
         for (size_t entityID = 0; entityID < MAX_ENTITIES; ++entityID) {
             if (world.entityExists(entityID)) {
                 const SpriteComponent& sprite = world.sprites[entityID];
